requirements.txt
streamlit
geopandas
folium
streamlit-folium
pandas
google-analytics-data
plotly
openpyxl
